requirements_test.txt
pytest==3.10.1
pytest-pep8==1.0.6
pytest-cov==2.6.1
pytest-flask==0.14.0
pytest-flask-sqlalchemy==1.0.2
coveralls==1.7.0
-r request_handler/requirements/production.txt
